+++ core-services/report/src/main/java/org/egov/controller/ReportController.java
@@ -1,5 +1,6 @@
 package org.egov.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.ReportApp;
 import org.egov.domain.model.MetaDataRequest;