+++ core-services/report/src/main/java/org/egov/domain/model/MetaDataRequest.java
@@ -1,7 +1,7 @@
 package org.egov.domain.model;
 
 import javax.validation.constraints.NotNull;
-
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.egov.common.contract.request.RequestInfo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;