+++ core-services/report/src/main/java/org/egov/domain/model/Response.java
@@ -14,7 +14,7 @@ public class Response {
         final String resMsgId = "uief87324"; // FIXME : Hard-coded
         final String msgId = requestInfo != null ? requestInfo.getMsgId() : "";
         final String responseStatus = success ? "successful" : "failed";
-        return new ResponseInfo(apiId, ver, ts, resMsgId, msgId, responseStatus);
+        return new ResponseInfo(apiId, ver, Long.valueOf(ts), resMsgId, msgId, responseStatus);
     }
 
 }