+++ core-services/report/src/main/java/org/egov/swagger/model/ColumnDetail.java
@@ -95,7 +95,9 @@ public class ColumnDetail extends ColumnDef {
         SINGLEVALUELISTAC("singlevaluelistac"),
         MULTIVALUELIST("multivaluelist"),
         MULTIVALUELISTAC("multivaluelistac"),
-        BOUNDARYLIST("boundarylist");
+        BOUNDARYLIST("boundarylist"),
+
+        STRINGARRAY("stringarray");
 
         private String value;